+++ source/extensions/omni.isaac.core/omni/isaac/core/utils/prims.py
@@ -38,6 +38,15 @@ def get_prim_at_path(prim_path: str, fabric: bool = False) -> typing.Union[Usd.P
 
     Returns:
         typing.Union[Usd.Prim, usdrt.Usd._Usd.Prim]: USD or Fabric Prim object at the given path in the current stage.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prims_utils.get_prim_at_path("/World/Cube")
+        Usd.Prim(</World/Cube>)
     """
     return get_current_stage(fabric=fabric).GetPrimAtPath(prim_path)
 
@@ -47,18 +56,77 @@ def is_prim_path_valid(prim_path: str, fabric: bool = False) -> bool:
 
     Args:
         prim_path (str): path of the prim in the stage
+        fabric (bool, optional): True for fabric stage and False for USD stage. Defaults to False.
 
     Returns:
         bool: True if the path points to a valid prim
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube
+        >>> prims_utils.is_prim_path_valid("/World/Cube")
+        True
+        >>> prims_utils.is_prim_path_valid("/World/Cube/")
+        False
+        >>> prims_utils.is_prim_path_valid("/World/Sphere")  # it doesn't exist
+        False
     """
     return get_prim_at_path(prim_path, fabric=fabric).IsValid()
 
 
-def get_prim_attribute_names(prim_path: str, fabric: bool = False):
+def get_prim_attribute_names(prim_path: str, fabric: bool = False) -> typing.List[str]:
+    """Get all the attribute names of a prim at the path
+
+    Args:
+        prim_path (str): path of the prim in the stage
+        fabric (bool, optional): True for fabric stage and False for USD stage. Defaults to False.
+
+    Raises:
+        Exception: If there is not a valid prim at the given path
+
+    Returns:
+        typing.List[str]: List of the prim attribute names
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prims_utils.get_prim_attribute_names("/World/Cube")
+        ['doubleSided', 'extent', 'orientation', 'primvars:displayColor', 'primvars:displayOpacity',
+         'purpose', 'size', 'visibility', 'xformOp:orient', 'xformOp:scale', 'xformOp:translate', 'xformOpOrder']
+    """
     return [attr.GetName() for attr in get_prim_at_path(prim_path=prim_path, fabric=fabric).GetAttributes()]
 
 
-def get_prim_attribute_value(prim_path: str, attribute_name: str, fabric: bool = False):
+def get_prim_attribute_value(prim_path: str, attribute_name: str, fabric: bool = False) -> typing.Any:
+    """Get a prim attribute value
+
+    Args:
+        prim_path (str): path of the prim in the stage
+        attribute_name (str): name of the attribute to get
+        fabric (bool, optional): True for fabric stage and False for USD stage. Defaults to False.
+
+    Raises:
+        Exception: If there is not a valid prim at the given path
+
+    Returns:
+        typing.Any: Prim attribute value
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prims_utils.get_prim_attribute_value("/World/Cube", attribute_name="size")
+        1.0
+    """
     attr = get_prim_at_path(prim_path=prim_path, fabric=fabric).GetAttribute(attribute_name)
     if fabric:
         type_name = str(attr.GetTypeName().GetAsString())
@@ -70,7 +138,24 @@ def get_prim_attribute_value(prim_path: str, attribute_name: str, fabric: bool =
         return attr.Get()
 
 
-def set_prim_attribute_value(prim_path: str, attribute_name: str, value: np.ndarray, fabric: bool = False):
+def set_prim_attribute_value(prim_path: str, attribute_name: str, value: typing.Any, fabric: bool = False):
+    """Set a prim attribute value
+
+    Args:
+        prim_path (str): path of the prim in the stage
+        attribute_name (str): name of the attribute to set
+        value (typing.Any): value to set the attribute to
+        fabric (bool, optional): True for fabric stage and False for USD stage. Defaults to False.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube. Set the Cube size to 5.0
+        >>> prims_utils.set_prim_attribute_value("/World/Cube", attribute_name="size", value=5.0)
+    """
     attr = get_prim_at_path(prim_path=prim_path, fabric=fabric).GetAttribute(attribute_name)
     if fabric:
         type_name = str(attr.GetTypeName().GetAsString())
@@ -87,35 +172,61 @@ def set_prim_attribute_value(prim_path: str, attribute_name: str, value: np.ndar
         attr.Set(value)
 
 
-def define_prim(prim_path: str, prim_type: str = "Xform", fabric=False) -> Usd.Prim:
+def define_prim(prim_path: str, prim_type: str = "Xform", fabric: bool = False) -> Usd.Prim:
     """Create a USD Prim at the given prim_path of type prim_type unless one already exists
+
+    .. note::
+
+        This method will create a prim of the specified type in the specified path.
+        To apply a transformation (position, orientation, scale), set attributes or
+        load an USD file while creating the prim use the ``create_prim`` function.
 
     Args:
         prim_path (str): path of the prim in the stage
         prim_type (str, optional): The type of the prim to create. Defaults to "Xform".
+        fabric (bool, optional): True for fabric stage and False for USD stage. Defaults to False.
 
     Raises:
         Exception: If there is already a prim at the prim_path
 
     Returns:
         Usd.Prim: The created USD prim.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prims_utils.define_prim("/World/Shapes", prim_type="Xform")
+        Usd.Prim(</World/Shapes>)
     """
     if is_prim_path_valid(prim_path, fabric=fabric):
         raise Exception("A prim already exists at prim path: {}".format(prim_path))
     return get_current_stage(fabric=fabric).DefinePrim(prim_path, prim_type)
 
 
-def get_prim_type_name(prim_path: str, fabric=False) -> str:
+def get_prim_type_name(prim_path: str, fabric: bool = False) -> str:
     """Get the TypeName of the USD Prim at the path if it is valid
 
     Args:
         prim_path (str): path of the prim in the stage
+        fabric (bool, optional): True for fabric stage and False for USD stage. Defaults to False.
 
     Raises:
         Exception: If there is not a valid prim at the given path
 
     Returns:
         str: The TypeName of the USD Prim at the path string
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prims_utils.get_prim_type_name("/World/Cube")
+        Cube
     """
     if not is_prim_path_valid(prim_path, fabric=fabric):
         raise Exception("A prim does not exist at prim path: {}".format(prim_path))
@@ -132,19 +243,43 @@ def move_prim(path_from: str, path_to: str) -> None:
     Args:
         path_from (str): Path of the USD Prim you wish to move
         path_to (str): Final destination of the prim
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube. Move the prim Cube outside the prim World
+        >>> prims_utils.move_prim("/World/Cube", "/Cube")
     """
     MovePrimCommand(path_from=path_from, path_to=path_to).do()
 
 
-def get_first_matching_child_prim(prim_path: str, predicate: typing.Callable[[str], bool], fabric=False) -> Usd.Prim:
+def get_first_matching_child_prim(
+    prim_path: str, predicate: typing.Callable[[str], bool], fabric: bool = False
+) -> Usd.Prim:
     """Recursively get the first USD Prim at the path string that passes the predicate function
 
     Args:
         prim_path (str): path of the prim in the stage
         predicate (typing.Callable[[str], bool]): Function to test the prims against
+        fabric (bool, optional): True for fabric stage and False for USD stage. Defaults to False.
 
     Returns:
          Usd.Prim: The first prim or child of the prim, as defined by GetChildren, that passes the predicate
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube, /World/Cube_01, /World/Cube_02.
+        >>> # Get the first child prim of type Cube
+        >>> predicate = lambda path: prims_utils.get_prim_type_name(path) == "Cube"
+        >>> prims_utils.get_first_matching_child_prim("/", predicate)
+        Usd.Prim(</World/Cube>)
     """
     prim = get_current_stage(fabric=fabric).GetPrimAtPath(prim_path)
     children_stack = [prim]
@@ -168,6 +303,17 @@ def get_first_matching_parent_prim(prim_path: str, predicate: typing.Callable[[s
 
     Returns:
         str: The first prim on the parent path, as defined by GetParent, that passes the predicate
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube. Get the first parent of Cube prim of type Xform
+        >>> predicate = lambda path: prims_utils.get_prim_type_name(path) == "Xform"
+        >>> prims_utils.get_first_matching_parent_prim("/World/Cube", predicate)
+        Usd.Prim(</World>)
     """
     current_prim_path = get_prim_path(get_prim_parent(get_prim_at_path(prim_path)))
     while not is_prim_root_path(current_prim_path):
@@ -190,6 +336,21 @@ def get_all_matching_child_prims(
 
     Returns:
         typing.List[Usd.Prim]: A list containing the root and children prims matching specified predicate.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # get all hidden prims
+        >>> predicate = lambda path: prims_utils.is_prim_hidden_in_stage(path)  # True if the prim at path is hidden
+        >>> prims_utils.get_all_matching_child_prims("/", predicate)
+        [Usd.Prim(</OmniverseKit_Persp>),
+         Usd.Prim(</OmniverseKit_Front>),
+         Usd.Prim(</OmniverseKit_Top>),
+         Usd.Prim(</OmniverseKit_Right>),
+         Usd.Prim(</Render>)]
     """
     prim = get_prim_at_path(prim_path)
     traversal_queue = [(prim, 0)]
@@ -213,6 +374,17 @@ def find_matching_prim_paths(prim_path_regex: str) -> typing.List[str]:
 
     Returns:
         typing.List[str]: List of prim paths that match input expression.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/env/Cube, /World/env_01/Cube, /World/env_02/Cube
+        >>> # get only the prim Cube paths from env_01 and env_02
+        >>> prims_utils.find_matching_prim_paths("/World/env_.*/Cube")
+        ['/World/env_01/Cube', '/World/env_02/Cube']
     """
     expressions_to_match = [prim_path_regex]
     result = []
@@ -246,6 +418,18 @@ def get_prim_children(prim: Usd.Prim) -> typing.List[Usd.Prim]:
 
     Returns:
         typing.List[Usd.Prim]: A list of the prim's children.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube, /World/Cube_01, /World/Cube_02.
+        >>> # Get all prims under the prim World
+        >>> prim = prims_utils.get_prim_at_path("/World")
+        >>> prims_utils.get_prim_children(prim)
+        [Usd.Prim(</World/Cube>), Usd.Prim(</World/Cube_01>), Usd.Prim(</World/Cube_02>)]
     """
     return prim.GetChildren()
 
@@ -258,6 +442,17 @@ def get_prim_parent(prim: Usd.Prim) -> Usd.Prim:
 
     Returns:
         Usd.Prim: The prim's parent returned from GetParent
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube. Get the prim Cube's parent
+        >>> prim = prims_utils.get_prim_at_path("/World/Cube")
+        >>> prims_utils.get_prim_parent(prim)
+        Usd.Prim(</World>)
     """
     return prim.GetParent()
 
@@ -271,6 +466,17 @@ def query_parent_path(prim_path: str, predicate: typing.Callable[[str], bool]) -
 
     Returns:
         bool: True if there is an ancestor that can pass the predicate, False otherwise
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube. Check is the prim Cube has an ancestor of type Xform
+        >>> predicate = lambda path: prims_utils.get_prim_type_name(path) == "Xform"
+        >>> prims_utils.query_parent_path("/World/Cube", predicate)
+        True
     """
     current_prim_path = get_prim_path(get_prim_parent(get_prim_at_path(prim_path)))
     while not is_prim_root_path(current_prim_path):
@@ -288,6 +494,21 @@ def is_prim_ancestral(prim_path: str) -> bool:
 
     Returns:
         True if prim is part of a referenced prim, false otherwise.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # /World/Cube is a prim created
+        >>> prims_utils.is_prim_ancestral("/World/Cube")
+        False
+        >>> # /World/panda is an USD file loaded (as reference) under that path
+        >>> prims_utils.is_prim_ancestral("/World/panda")
+        False
+        >>> prims_utils.is_prim_ancestral("/World/panda/panda_link0")
+        True
     """
     return omni.usd.check_ancestral(get_prim_at_path(prim_path))
 
@@ -300,6 +521,20 @@ def is_prim_root_path(prim_path: str) -> bool:
 
     Returns:
         True if the prim path is "/", False otherwise
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube
+        >>> prims_utils.is_prim_root_path("/")
+        True
+        >>> prims_utils.is_prim_root_path("/World")
+        False
+        >>> prims_utils.is_prim_root_path("/World/Cube")
+        False
     """
     if prim_path == "/":
         return True
@@ -310,11 +545,30 @@ def is_prim_root_path(prim_path: str) -> bool:
 def is_prim_no_delete(prim_path: str) -> bool:
     """Checks whether a prim can be deleted or not from USD stage.
 
+    .. note ::
+
+        This function checks for the ``no_delete`` prim metadata. A prim with this
+        metadata set to True cannot be deleted by using the edit menu, the context menu,
+        or by calling the ``delete_prim`` function, for example.
+
     Args:
         prim_path (str): The path to the USD prim.
 
     Returns:
         True if prim cannot be deleted, False if it can
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # prim without the 'no_delete' metadata
+        >>> prims_utils.is_prim_no_delete("/World/Cube")
+        None
+        >>> # prim with the 'no_delete' metadata set to True
+        >>> prims_utils.is_prim_no_delete("/World/Cube")
+        True
     """
     return get_prim_at_path(prim_path).GetMetadata("no_delete")
 
@@ -322,14 +576,29 @@ def is_prim_no_delete(prim_path: str) -> bool:
 def is_prim_hidden_in_stage(prim_path: str) -> bool:
     """Checks if the prim is hidden in the USd stage or not.
 
+    .. warning ::
+
+        This function checks for the ``hide_in_stage_window`` prim metadata.
+        This metadata is not related to the visibility of the prim.
+
     Args:
         prim_path (str): The path to the USD prim.
 
-    Note:
-        This is not related to the prim visibility.
-
     Returns:
         True if prim is hidden from stage window, False if not hidden.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # prim without the 'hide_in_stage_window' metadata
+        >>> prims_utils.is_prim_hidden_in_stage("/World/Cube")
+        None
+        >>> # prim with the 'hide_in_stage_window' metadata set to True
+        >>> prims_utils.is_prim_hidden_in_stage("/World/Cube")
+        True
     """
     return get_prim_at_path(prim_path).GetMetadata("hide_in_stage_window")
 
@@ -342,6 +611,16 @@ def get_prim_path(prim: Usd.Prim) -> str:
 
     Returns:
         str: The path to the input prim.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prim = prims_utils.get_prim_at_path("/World/Cube")  # Usd.Prim(</World/Cube>)
+        >>> prims_utils.get_prim_path(prim)
+        /World/Cube
     """
     if isinstance(prim, Usd.Prim):
         return prim.GetPath().pathString
@@ -352,11 +631,23 @@ def get_prim_path(prim: Usd.Prim) -> str:
 def set_prim_visibility(prim: Usd.Prim, visible: bool) -> None:
     """Sets the visibility of the prim in the opened stage.
 
-    The method does this through the USD API.
+    .. note::
+
+        The method does this through the USD API.
 
     Args:
         prim (Usd.Prim): the USD prim
         visible (bool): flag to set the visibility of the usd prim in stage.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube. Make the Cube not visible
+        >>> prim = prims_utils.get_prim_at_path("/World/Cube")
+        >>> prims_utils.set_prim_visibility(prim, False)
     """
     imageable = UsdGeom.Imageable(prim)
     if visible:
@@ -398,6 +689,34 @@ def create_prim(
 
     Returns:
         Usd.Prim: The created USD prim.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import numpy as np
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # create a cube (/World/Cube) of size 2 centered at (1.0, 0.5, 0.0)
+        >>> prims_utils.create_prim(
+        ...     prim_path="/World/Cube",
+        ...     prim_type="Cube",
+        ...     position=np.array([1.0, 0.5, 0.0]),
+        ...     attributes={"size": 2.0}
+        ... )
+        Usd.Prim(</World/Cube>)
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # load an USD file (franka.usd) to the stage under the path /World/panda
+        >>> prims_utils.create_prim(
+        ...     prim_path="/World/panda",
+        ...     prim_type="Xform",
+        ...     usd_path="/home/<user>/Documents/Assets/Robots/Franka/franka.usd"
+        ... )
+        Usd.Prim(</World/panda>)
     """
     # Note: Imported here to prevent cyclic dependency in the module.
     from omni.isaac.core.prims.xform_prim import XFormPrim
@@ -427,6 +746,14 @@ def delete_prim(prim_path: str) -> None:
 
     Args:
         prim_path (str): path of the prim in the stage
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prims_utils.delete_prim("/World/Cube")
     """
     DeletePrimsCommand([prim_path]).do()
 
@@ -440,6 +767,15 @@ def get_prim_property(prim_path: str, property_name: str) -> typing.Any:
 
     Returns:
         typing.Any: The attribute if it exists, None otherwise
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prims_utils.get_prim_property("/World/Cube", property_name="size")
+        1.0
     """
     prim = get_prim_at_path(prim_path=prim_path)
     return prim.GetAttribute(property_name).Get()
@@ -452,15 +788,24 @@ def set_prim_property(prim_path: str, property_name: str, property_value: typing
         prim_path (str): path of the prim in the stage
         property_name (str): name of the attribute to set
         property_value (typing.Any): value to set the attribute to
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube. Set the Cube size to 5.0
+        >>> prims_utils.set_prim_property("/World/Cube", property_name="size", property_value=5.0)
     """
     prim = get_prim_at_path(prim_path=prim_path)
     prim.GetAttribute(property_name).Set(property_value)
 
 
 def get_prim_object_type(prim_path: str) -> typing.Union[str, None]:
-    """Get the dynamic control Ooject type of the USD Prim at the given path.
+    """Get the dynamic control object type of the USD Prim at the given path.
 
-    If the prim at the path is of Dynamic Control type--i.e. rigid_body, joint, dof, articulation, attractor, d6joint,
+    If the prim at the path is of Dynamic Control type e.g.: rigid_body, joint, dof, articulation, attractor, d6joint,
     then the correspodning string returned. If is an Xformable prim, then "xform" is returned. Otherwise None
     is returned.
 
@@ -472,6 +817,15 @@ def get_prim_object_type(prim_path: str) -> typing.Union[str, None]:
 
     Returns:
         str: String corresponding to the object type.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prims_utils.get_prim_object_type("/World/Cube")
+        xform
     """
     dc_interface = _dynamic_control.acquire_dynamic_control_interface()
     object_type = dc_interface.peek_object_type(prim_path)
@@ -506,6 +860,19 @@ def is_prim_non_root_articulation_link(prim_path: str) -> bool:
     Returns:
         bool: True if the prim path corresponds to a link in an articulation which is a non root link
               and can't have a transformation applied to it.
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # /World/panda contains the prim tree for the Franka panda robot.
+        >>> # The prim on this path has the Physics Articulation Root property applied
+        >>> prims_utils.is_prim_non_root_articulation_link("/World/panda")
+        False
+        >>> prims_utils.is_prim_non_root_articulation_link("/World/panda/panda_link0")
+        True
     """
     parent_articulation_root = get_first_matching_parent_prim(
         prim_path=prim_path, predicate=lambda a: get_prim_at_path(a).HasAPI(UsdPhysics.ArticulationRootAPI)
@@ -534,21 +901,49 @@ def is_prim_non_root_articulation_link(prim_path: str) -> bool:
 
 
 def set_prim_hide_in_stage_window(prim: Usd.Prim, hide: bool):
-    """set hide_in_stage_window metadata for prim
+    """Set ``hide_in_stage_window`` metadata for a prim
+
+    .. warning ::
+
+        This metadata is unrelated to the visibility of the prim.
+        Use the ``set_prim_visibility`` function for the latter purpose
 
     Args:
         prim (Usd.Prim): Prim to set
         hide (bool): True to hide in stage window, false to show
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prim = prims_utils.get_prim_at_path("/World/Cube")
+        >>> prims_utils.set_prim_hide_in_stage_window(prim, True)
     """
     prim.SetMetadata("hide_in_stage_window", hide)
 
 
 def set_prim_no_delete(prim: Usd.Prim, no_delete: bool):
-    """set no_delete metadata for prim
+    """Set ``no_delete`` metadata for a prim
+
+    .. note ::
+
+        A prim with this metadata set to True cannot be deleted by using the edit menu,
+        the context menu, or by calling the ``delete_prim`` function, for example.
 
     Args:
         prim (Usd.Prim): Prim to set
         no_delete (bool):True to make prim undeletable in stage window, false to allow deletion
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> prim = prims_utils.get_prim_at_path("/World/Cube")
+        >>> prims_utils.set_prim_no_delete(prim, True)
     """
     prim.SetMetadata("no_delete", no_delete)
 
@@ -560,6 +955,18 @@ def set_targets(prim: Usd.Prim, attribute: str, target_prim_paths: list):
         prim (Usd.Prim): Prim to create and set attribute on
         attribute (str): Relationship attribute to create
         target_prim_paths (list): list of targets to set
+
+    Example:
+
+    .. code-block:: python
+
+        >>> import omni.isaac.core.utils.prims as prims_utils
+        >>>
+        >>> # given the stage: /World/Cube, /World/Cube_01, /World/Cube_02.
+        >>> # Set each prim Cube to the relationship targetPrim of the prim World
+        >>> prim = prims_utils.get_prim_at_path("/World")
+        >>> targets = ["/World/Cube", "/World/Cube_01", "/World/Cube_02"]
+        >>> prims_utils.set_targets(prim, "targetPrim", targets)
     """
     try:
         input_rel = prim.CreateRelationship(attribute)